backend/requirements.txt
fastapi[all]
haystack-ai
sentence-transformers>=2.2.0
huggingface_hub>=0.22.0
markdown-it-py
mdit_plain
pypdf
gdown
